Use full width of the terminal
Long descriptions are not easy to read. Now we are getting the width of
the terminal, if possible and using all of it. We remove TABLE_PADDING of 4
since the borders and the padding of the tablewriter otherwise make the
text overlap.

If we can't get the width of the terminal we fallback to width of 80.